Move credentials.go to xds/bootstrap
