+++ internal/xds/bootstrap/bootstrap_test.go
@@ -1018,43 +1018,3 @@ func TestDefaultBundles(t *testing.T) {
 		})
 	}
 }
-
-func TestCredsBuilders(t *testing.T) {
-	tests := []struct {
-		typename string
-		builder  bootstrap.Credentials
-	}{
-		{"google_default", &googleDefaultCredsBuilder{}},
-		{"insecure", &insecureCredsBuilder{}},
-		{"tls", &tlsCredsBuilder{}},
-	}
-
-	for _, test := range tests {
-		t.Run(test.typename, func(t *testing.T) {
-			if got, want := test.builder.Name(), test.typename; got != want {
-				t.Errorf("%T.Name = %v, want %v", test.builder, got, want)
-			}
-
-			_, stop, err := test.builder.Build(nil)
-			if err != nil {
-				t.Fatalf("%T.Build failed: %v", test.builder, err)
-			}
-			stop()
-		})
-	}
-}
-
-func TestTlsCredsBuilder(t *testing.T) {
-	tls := &tlsCredsBuilder{}
-	_, stop, err := tls.Build(json.RawMessage(`{}`))
-	if err != nil {
-		t.Fatalf("tls.Build() failed with error %s when expected to succeed", err)
-	}
-	stop()
-
-	if _, stop, err := tls.Build(json.RawMessage(`{"ca_certificate_file":"/ca_certificates.pem","refresh_interval": "asdf"}`)); err == nil {
-		t.Errorf("tls.Build() succeeded with an invalid refresh interval, when expected to fail")
-		stop()
-	}
-	// package internal/xdsclient/tlscreds has tests for config validity.
-}
